requirements.txt
crewai==0.28.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
google-generativeai==0.3.2
python-dotenv==1.0.0
click==8.1.7
pandas==2.1.1
matplotlib==3.8.0
tqdm==4.66.1 